Add files via upload
This was simply put, a very easy task to do in Lua. All it did was, when the user typed "/help" in the chat, the bot would display a greeting and a list of extra help commands they could type in regarding different topics. These were based off of the FAQ that was in the server's Discord. A popular one was how does one obtain a K9. Displayed was an option to hear how to obtain it. The user would type in "help-k9" or something similiar and it would display in chat how to go about getting one. It was very helpful for the newer players.