Add files via upload
Similar to the previous help project, I began work on a virtual internet that would later become something massive. It contains a lot of ingame lore that is more suitable for those who've played GTA and are familiar with the various parody sites. It was an unfinished project that may never see the light of day.